pyproject.toml
[project]
name = "ai-meeting-notes-agent"
version = "0.4.0"
requires-python = ">=3.10"
dependencies = [
    "assemblyai==0.17.0",
    "moviepy==1.0.3",
    "python-dotenv==1.0.0",
    "tqdm==4.66.1",
    "zhconv==1.4.3",
    "python-telegram-bot>=21.0",
    "openai>=1.0.0",
    "claude-agent-sdk>=0.1.0",
    "boto3>=1.34.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
]
